+++ test/datastore.spec.js
@@ -20,13 +20,13 @@ const domainDir = join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com')
 
 test.group('Datastore', (group) => {
   group.afterEach(async () => {
-    await fs.remove(domainDir)
+    await fs.remove(join(__dirname, '..', 'sites'))
   })
 
   test('raise error when versionNo or fileName is missing when saving doc', async (assert) => {
     assert.plan(1)
 
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
 
     try {
       await store.saveDoc()
@@ -36,7 +36,7 @@ test.group('Datastore', (group) => {
   })
 
   test('save markdown json and it\'s meta data to the disk', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     await store.saveDoc('1.0.0', 'foo.md', {
@@ -77,7 +77,7 @@ test.group('Datastore', (group) => {
   })
 
   test('pull title from json when is not inside meta data', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const nodes = {
@@ -85,8 +85,8 @@ test.group('Datastore', (group) => {
       children: [
         {
           t: 'elem',
-          tag: 'dimerTitle',
-          child: [{ t: 'text', value: 'Hello world' }]
+          tag: 'dimertitle',
+          children: [{ t: 'text', value: 'Hello world' }]
         }
       ]
     }
@@ -124,7 +124,7 @@ test.group('Datastore', (group) => {
   })
 
   test('set category to root when is not inside meta data', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const nodes = {
@@ -166,7 +166,7 @@ test.group('Datastore', (group) => {
   })
 
   test('sync versions', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const { added, removed } = await store.syncVersions([
@@ -209,7 +209,7 @@ test.group('Datastore', (group) => {
   })
 
   test('sync versions with existing content should work fine', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const nodes = {
@@ -278,7 +278,7 @@ test.group('Datastore', (group) => {
   })
 
   test('sync versions should remove the one\'s not inside the array', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const nodes = {
@@ -341,7 +341,7 @@ test.group('Datastore', (group) => {
   })
 
   test('sync versions should add the one\'s not already exists', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const nodes = {
@@ -400,7 +400,7 @@ test.group('Datastore', (group) => {
   })
 
   test('sync versions should remove the content for one\'s not inside the array', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const nodes = {
@@ -462,7 +462,7 @@ test.group('Datastore', (group) => {
   })
 
   test('remove a given doc', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const nodes = {
@@ -498,7 +498,7 @@ test.group('Datastore', (group) => {
   })
 
   test('skip when removing doc for a non-existing version', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const nodes = {
@@ -541,7 +541,7 @@ test.group('Datastore', (group) => {
   })
 
   test('return an array of versions', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const nodes = {
@@ -575,7 +575,7 @@ test.group('Datastore', (group) => {
   })
 
   test('get docs for an array of version', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const nodes = {
@@ -606,7 +606,7 @@ test.group('Datastore', (group) => {
   })
 
   test('order versions by jsonPath', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const nodes = {
@@ -649,7 +649,7 @@ test.group('Datastore', (group) => {
   })
 
   test('load content for docs', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const nodes = {
@@ -694,7 +694,7 @@ test.group('Datastore', (group) => {
   })
 
   test('limit docs', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const nodes = {
@@ -732,7 +732,7 @@ test.group('Datastore', (group) => {
   })
 
   test('get a single doc', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const nodes = {
@@ -756,8 +756,27 @@ test.group('Datastore', (group) => {
     })
   })
 
+  test('return null when single doc is missing', async (assert) => {
+    const store = new Datastore(domainDir)
+    await store.db.load()
+
+    const nodes = {
+      type: 'root',
+      children: [{}]
+    }
+
+    await store.saveDoc('1.0.0', 'foo.md', {
+      title: 'Hello world',
+      permalink: '/hello',
+      content: nodes
+    })
+
+    const doc = await store.getDoc('1.0.0', 'bar.json')
+    assert.isNull(doc)
+  })
+
   test('get a single doc by permalink', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const nodes = {
@@ -781,8 +800,27 @@ test.group('Datastore', (group) => {
     })
   })
 
+  test('return null if doc is missing', async (assert) => {
+    const store = new Datastore(domainDir)
+    await store.db.load()
+
+    const nodes = {
+      type: 'root',
+      children: [{}]
+    }
+
+    await store.saveDoc('1.0.0', 'foo.md', {
+      title: 'Hello world',
+      permalink: '/hello',
+      content: nodes
+    })
+
+    const doc = await store.getDocByPermalink('1.0.0', 'foo')
+    assert.isNull(doc)
+  })
+
   test('index docs for a given version', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const template = dedent`
@@ -830,7 +868,7 @@ test.group('Datastore', (group) => {
   })
 
   test('search docs for a given term', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const template = dedent`
@@ -881,7 +919,7 @@ test.group('Datastore', (group) => {
   test('raise error when permalink is same', async (assert) => {
     assert.plan(2)
 
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     await store.saveDoc('1.0.0', 'foo.md', {
@@ -918,7 +956,7 @@ test.group('Datastore', (group) => {
   test('work fine when updating doc with same permalink', async (assert) => {
     assert.plan(2)
 
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const nodes = {
@@ -968,7 +1006,7 @@ test.group('Datastore', (group) => {
   })
 
   test('indexing version should start from scratch', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const template = dedent`
@@ -996,7 +1034,7 @@ test.group('Datastore', (group) => {
   })
 
   test('get the permalink of the actual doc when redirected permalink is accessed', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const template = dedent`
@@ -1020,7 +1058,7 @@ test.group('Datastore', (group) => {
   })
 
   test('get a single doc by permalink by normalizing slashes', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const nodes = {
@@ -1045,7 +1083,7 @@ test.group('Datastore', (group) => {
   })
 
   test('load store from a blank slate', async (assert) => {
-    const store = new Datastore(join(__dirname, '..', 'sites', 'adonisjs.dimerapp.com'))
+    const store = new Datastore(domainDir)
     await store.db.load()
 
     const nodes = {